+++ src/imio/actionspanel/browser/views.py
@@ -1,8 +1,11 @@
 import logging
+logger = logging.getLogger('imio.actionspanel')
 
 from appy.gen import No
 
 from Acquisition import aq_base
+
+from AccessControl import Unauthorized
 
 from zope.i18n import translate
 
@@ -17,6 +20,9 @@ from Products.DCWorkflow.Expression import createExprContext
 from Products.DCWorkflow.Transitions import TRIGGER_USER_ACTION
 
 from imio.actionspanel import ActionsPanelMessageFactory as _
+from imio.actionspanel.interfaces import IContentDeletable
+from imio.actionspanel.utils import unrestrictedRemoveGivenObject
+
 
 class ActionsPanelView(BrowserView):
     """
@@ -32,24 +38,24 @@ class ActionsPanelView(BrowserView):
             self.request.set('imio.actionspanel_member_cachekey', self.member)
         self.SECTIONS_TO_RENDER = ('renderTransitions',
                                    'renderEdit',
-                                   'renderActions',
-                                   'renderAllowedContentTypes'
-                                   )
+                                   'renderOwnDelete',
+                                   'renderActions', )
         # portal_actions.object_buttons action ids not to keep
         # every actions will be kept except actions listed here
         self.IGNORABLE_ACTIONS = ()
+
         # portal_actions.object_buttons action ids to keep
         # if you define some here, only these actions will be kept
         self.ACCEPTABLE_ACTIONS = ()
 
-    def render(self,
-               useIcons=True,
-               showTransitions=True,
-               appendTypeNameToTransitionLabel=False,
-               showEdit=True,
-               showActions=True,
-               showAllowedContentTypes=True,
-               **kwargs):
+    def __call__(self,
+                 useIcons=True,
+                 showTransitions=True,
+                 appendTypeNameToTransitionLabel=False,
+                 showEdit=True,
+                 showOwnDelete=True,
+                 showActions=True,
+                 **kwargs):
         """
           Master method that will render the content.
           This is not supposed to be overrided.
@@ -58,8 +64,11 @@ class ActionsPanelView(BrowserView):
         self.showTransitions = showTransitions
         self.appendTypeNameToTransitionLabel = appendTypeNameToTransitionLabel
         self.showEdit = showEdit
+        self.showOwnDelete = showOwnDelete
+        # if we manage our own delete, do not use Plone default one
+        if self.showOwnDelete and not 'delete' in self.IGNORABLE_ACTIONS:
+            self.IGNORABLE_ACTIONS = self.IGNORABLE_ACTIONS + ('delete', )
         self.showActions = showActions
-        self.showAllowedContentTypes = showAllowedContentTypes
         self.kwargs = kwargs
         self.hasActions = False
         return self.index()
@@ -93,19 +102,22 @@ class ActionsPanelView(BrowserView):
             return ViewPageTemplateFile("actions_panel_edit.pt")(self)
         return ''
 
+    def renderOwnDelete(self):
+        """
+          Render our own version of the 'delete' action.
+        """
+        if self.showOwnDelete and \
+           self.member.has_permission('Delete objects', self.context) and \
+           IContentDeletable(self.context).mayDelete():
+            return ViewPageTemplateFile("actions_panel_own_delete.pt")(self)
+        return ''
+
     def renderActions(self):
         """
           Render actions coming from portal_actions.object_buttons and available on the context.
         """
         if self.showActions:
             return ViewPageTemplateFile("actions_panel_actions.pt")(self)
-
-    def renderAllowedContentTypes(self):
-        """
-          Render allowed_content_types coming from portal_type.
-        """
-        if self.showAllowedContentTypes:
-            return ViewPageTemplateFile("actions_panel_allowed_content_types.pt")(self)
 
     def mayEdit(self):
         """
@@ -244,36 +256,6 @@ class ActionsPanelView(BrowserView):
             return res
         return 1
 
-    def allowed_content_types(self):
-        """Return content types allowed"""
-
-        actions = []
-
-        types_tool = getToolByName(self, 'portal_types')
-        portal_type = types_tool.get(self.context.portal_type)
-        allowed_content_types = portal_type.allowed_content_types
-        import ipdb; ipdb.set_trace()
-        for content_type in allowed_content_types:
-            portal_type = types_tool.get(content_type)
-            add_permission = portal_type.add_permission
-            if checkPermission(add_permission, self.context):
-                url = '{}/++add++{}'.format(
-                    self.context.absolute_url(),
-                    content_type
-                )
-                action = '<a name=add_{} href={} class={} >\
-                    add {}\
-                    </a>'.format(
-                        content_type,
-                        url,
-                        "apButton apButtonAction",
-                        content_type
-                    )
-                actions.append(action)
-        actions = ''.join(actions)
-        actions = '<span>{}</span>'.format(actions)
-        return actions
-
     def listObjectButtonsActions(self):
         """
           Return a list of object_buttons actions coming from portal_actions.
@@ -329,7 +311,6 @@ class ActionsPanelView(BrowserView):
             # fail silently if the user triggered a transition he could not
             # this avoid WorkflowException error in the UI if a user double-click on an icon
             # triggering a workflow transition
-            logger = logging.getLogger('imio.actionspanel')
             logger.info("WorkflowException in imio.actionspanel.triggerTransition, the user '%s' "
                         "tried to trigger the transition '%s' but he could not.  Double click in the UI?" %
                         (self.member.getId(), self.request.get('transition')))
@@ -370,3 +351,68 @@ class ActionsPanelView(BrowserView):
         """
         urlBack = self.request['HTTP_REFERER']
         return self.request.RESPONSE.redirect(urlBack)
+
+
+class DeleteGivenUidView(BrowserView):
+    """
+      View that ease deletion of elements by not checking the 'Delete objects' permission on parent
+      but only on the object to delete itself.
+      Callable using self.portal.restrictedTraverse('@@delete_givenuid)(object_to_delete.UID()) in the code
+      and using classic traverse in a url : http://nohost/plonesite/delete_givenuid?object_uid=anUID
+    """
+    def __init__(self, context, request):
+        super(DeleteGivenUidView, self).__init__(context, request)
+        self.context = context
+        self.request = request
+        self.portal = getToolByName(self.context, 'portal_url').getPortalObject()
+
+    def __call__(self, object_uid):
+        # Get the object to delete
+        # try to get it from the portal_catalog
+        catalog_brains = self.context.portal_catalog(UID=object_uid)
+        # if not found, try to get it from the uid_catalog
+        if not catalog_brains:
+            catalog_brains = self.context.uid_catalog(UID=object_uid)
+        # if nto found at all, raise
+        if not catalog_brains:
+            raise KeyError('The given uid could not be found!')
+        obj = catalog_brains[0].getObject()
+        objectUrl = obj.absolute_url()
+
+        # we use an adapter to manage if we may delete the object
+        # that checks if the user has the 'Delete objects' permission
+        # on the content by default but that could be overrided
+        self.member = getToolByName(self.context, 'portal_membership').getAuthenticatedMember()
+        if self.member.has_permission("Delete objects", obj) and IContentDeletable(obj).mayDelete():
+            msg = {'message': _('object_deleted'),
+                   'type': 'info'}
+            # remove the object
+            # just manage BeforeDeleteException because we rise it ourselves
+            from OFS.ObjectManager import BeforeDeleteException
+            try:
+                unrestrictedRemoveGivenObject(obj)
+            except BeforeDeleteException, exc:
+                msg = {'message': exc.message,
+                       'type': 'error'}
+        else:
+            # as the action calling delete_givenuid is already protected by the chek
+            # made in the 'if' here above, if we arrive here it is that user is doing
+            # something wrong, we raise Unauthorized
+            raise Unauthorized
+
+        # Redirect the user to the correct page and display the correct message.
+        refererUrl = self.request['HTTP_REFERER']
+        if not refererUrl.startswith(objectUrl):
+            backURL = refererUrl
+        else:
+            backURL = self._computeBackURL()
+        self.portal.plone_utils.addPortalMessage(**msg)
+        return self.request.RESPONSE.redirect(backURL)
+
+    def _computeBackURL(self):
+        '''This is made to be overriden...'''
+        # find a parent the current user may access
+        parent = self.context.getParentNode()
+        while (not self.member.has_permission('View', parent) and not parent.meta_type == 'Plone Site'):
+            parent = parent.getParentNode()
+        return parent.absolute_url()